requirements.txt
numpy
streamlit
tensorflow
google-generativeai